requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
pytz==2023.3
flask==3.0.2
gunicorn==21.2.0
